+++ packages/web/src/pages/sign-up-page/pages/CreateEmailPage/CreateEmailPage.tsx
@@ -16,19 +16,7 @@ import { SIGN_IN_PAGE, SIGN_UP_PASSWORD_PAGE } from 'utils/route'
 import styles from './CreateEmailPage.module.css'
 import { CreateEmailPageDesktop } from './CreateEmailPageDesktop'
 import { CreateEmailPageMobile } from './CreateEmailPageMobile'
-
-const messages = {
-  title: 'Sign Up For Audius',
-  emailLabel: 'Email',
-  signUp: 'Sign Up Free',
-  haveAccount: 'Already have an account?',
-  signIn: 'Sign In',
-  subHeader:
-    'Join the revolution in music streaming! Discover, connect, and create on Audius.',
-  socialsDividerText: 'Or, get started with one of your socials',
-  invalidEmail: 'Please enter a valid email.',
-  unknownError: 'Unknown error occurred.'
-}
+import { messages } from './messages'
 
 type SignUpEmailValues = {
   email: string
