+++ packages/web/src/pages/sign-up-page/pages/CreateEmailPage/CreateEmailPageMobile.tsx
@@ -20,23 +20,7 @@ import {
 import { SIGN_IN_PAGE } from 'utils/route'
 
 import styles from './CreateEmailPage.module.css'
-
-const messages = {
-  title: 'Sign Up For Audius',
-  emailLabel: 'Email',
-  signUp: 'Sign Up Free',
-  haveAccount: 'Already have an account?',
-  signIn: 'Sign In',
-  subHeader: (
-    <>
-      Join the revolution in music streaming! <br /> Discover, connect, and
-      create on Audius.
-    </>
-  ),
-  socialsDividerText: 'Or, get started with one of your socials',
-  invalidEmail: 'Please enter a valid email.',
-  unknownError: 'Unknown error occurred.'
-}
+import { messages } from './messages'
 
 export const CreateEmailPageMobile = ({
   isSubmitting
